refactor: move truncation into shared api key lib
