+++ packages/config/src/index.ts
@@ -28,4 +28,3 @@ export { ConfigJson, zoomLevelsFromWmts } from './json/json.config.js';
 export { standardizeLayerName } from './json/name.convertor.js';
 export { base58, isBase58 } from './base58.js';
 export { sha256base58, ensureBase58 } from './base58.node.js';
-export { truncateApiKey } from './api.truncate.js';
