+++ packages/lambda-tiler/src/util/validate.ts
@@ -1,9 +1,8 @@
 import { ImageFormat, Projection, TileMatrixSet, TileMatrixSets, VectorFormat } from '@basemaps/geo';
-import { Const, isValidApiKey } from '@basemaps/shared';
+import { Const, isValidApiKey, truncateApiKey } from '@basemaps/shared';
 import { getImageFormat } from '@basemaps/tiler';
 import { LambdaHttpRequest, LambdaHttpResponse } from '@linzjs/lambda';
 import { TileXyzGet } from '../routes/tile.xyz';
-import { truncateApiKey } from '@basemaps/config';
 
 export interface TileXyz {
   tile: { x: number; y: number; z: number };
